requirements.txt
discord.py
python-dotenv
asyncpg
redis
qdrant-client
google-generativeai
openai
psutil
